+++ src/api/__tests__/APIFactory-test.js
@@ -3,7 +3,7 @@
 import APIFactory from '../APIFactory';
 import Cache from '../../util/Cache';
 import ChunkedUploadAPI from '../uploads/MultiputUpload';
-import PlainUploadAPI from '../PlainUpload';
+import PlainUploadAPI from '../uploads/PlainUpload';
 import FolderAPI from '../Folder';
 import FileAPI from '../File';
 import WebLinkAPI from '../WebLink';
