+++ src/api/uploads/BaseMultiput.js
@@ -3,7 +3,7 @@
  * @file Multiput upload base class
  * @author Box
  */
-import Base from '../Base';
+import BaseUpload from './BaseUpload';
 import type { MultiputConfig, Options } from '../../flowTypes';
 
 const DEFAULT_MULTIPUT_CONFIG: MultiputConfig = {
@@ -16,7 +16,7 @@ const DEFAULT_MULTIPUT_CONFIG: MultiputConfig = {
     retries: 5 // How many times to retry requests such as upload part or commit. Note that total number of attempts will be retries + 1 in worst case where all attempts fail.
 };
 
-class BaseMultiput extends Base {
+class BaseMultiput extends BaseUpload {
     config: MultiputConfig;
     sessionEndpoints: Object;
