+++ src/api/uploads/__tests__/PlainUpload-test.js
@@ -1,167 +1,12 @@
 import PlainUpload from '../PlainUpload';
 
 let upload;
-let clock;
 let file;
 
-describe('api/PlainUpload', () => {
+describe('api/uploads/PlainUpload', () => {
     beforeEach(() => {
         upload = new PlainUpload({
             token: '123'
-        });
-    });
-
-    describe('uploadErrorHandler()', () => {
-        beforeEach(() => {
-            clock = jest.useFakeTimers();
-            file = {
-                name: 'foo'
-            };
-            upload.file = file;
-        });
-
-        afterEach(() => {
-            jest.clearAllTimers();
-        });
-
-        test('should not do anything if API is destroyed', () => {
-            upload.isDestroyed = jest.fn().mockReturnValueOnce(true);
-            upload.makePreflightRequest = jest.fn();
-            upload.uploadErrorHandler(new Error());
-            expect(upload.makePreflightRequest).not.toHaveBeenCalled();
-        });
-
-        test('should overwrite file on 409 if overwrite property is true', () => {
-            const fileId = '123';
-            upload.overwrite = true;
-            upload.makePreflightRequest = jest.fn();
-
-            upload.uploadErrorHandler({
-                status: 409,
-                context_info: {
-                    conflicts: {
-                        id: fileId
-                    }
-                }
-            });
-
-            expect(upload.makePreflightRequest).toHaveBeenCalledWith({
-                fileId,
-                fileName: file.name
-            });
-        });
-
-        test('should append timestamp and re-upload on 409 if overwrite property is false', () => {
-            upload.file.name = 'foo.bar';
-            upload.overwrite = false;
-            Date.now = jest.fn().mockReturnValueOnce('1969-07-16');
-
-            upload.makePreflightRequest = jest.fn();
-
-            upload.uploadErrorHandler({
-                status: 409
-            });
-
-            expect(upload.makePreflightRequest).toHaveBeenCalledWith({
-                fileName: 'foo-1969-07-16.bar'
-            });
-        });
-
-        test('should retry on 429 status after default retry interval', () => {
-            const retryAfterMs = 3000;
-
-            upload.makePreflightRequest = jest.fn();
-
-            upload.uploadErrorHandler({
-                status: 429
-            });
-            clock.runTimersToTime(retryAfterMs + 1);
-
-            expect(upload.makePreflightRequest).toHaveBeenCalledWith({
-                fileName: file.name
-            });
-        });
-
-        test('should retry on too_many_requests code after default retry interval', () => {
-            const retryAfterMs = 3000;
-
-            upload.makePreflightRequest = jest.fn();
-
-            upload.uploadErrorHandler({
-                code: 'too_many_requests'
-            });
-            clock.runTimersToTime(retryAfterMs + 1);
-
-            expect(upload.makePreflightRequest).toHaveBeenCalledWith({
-                fileName: file.name
-            });
-        });
-
-        test('should retry after interval defined in response header', () => {
-            const retryAfterMs = 1000;
-
-            upload.makePreflightRequest = jest.fn();
-            const getMock = jest.fn().mockReturnValueOnce(`${retryAfterMs / 1000}`);
-
-            upload.uploadErrorHandler({
-                code: 'too_many_requests',
-                headers: {
-                    get: getMock
-                }
-            });
-            clock.runTimersToTime(retryAfterMs + 1);
-
-            expect(upload.makePreflightRequest).toHaveBeenCalledWith({
-                fileName: file.name
-            });
-
-            expect(getMock).toHaveBeenCalledWith('Retry-After');
-        });
-
-        test('should call the error callback if error has a status that is not 409 or 429', () => {
-            const error = new Error();
-            error.status = 500;
-
-            upload.errorCallback = jest.fn();
-            upload.uploadErrorHandler(error, () => {});
-            expect(upload.errorCallback).toHaveBeenCalledWith(error);
-        });
-
-        test('should call the error callback if error message is from CORS', () => {
-            const error = new Error();
-            error.message = 'Failed to fetch';
-
-            upload.errorCallback = jest.fn();
-            upload.uploadErrorHandler(error, () => {});
-            expect(upload.errorCallback).toHaveBeenCalledWith(error);
-        });
-
-        test('should retry after default interval for other errors', () => {
-            const retryAfterMs = 1000;
-            const error = new Error('Some other error');
-            upload.makePreflightRequest = jest.fn();
-
-            upload.uploadErrorHandler(error);
-            clock.runTimersToTime(retryAfterMs + 1);
-
-            expect(upload.makePreflightRequest).toHaveBeenCalledWith({
-                fileName: file.name
-            });
-        });
-
-        test('should not retry before exponential backoff interval for other errors', () => {
-            const MS_IN_S = 1000;
-            const retryCount = 3;
-            const error = new Error('Some other error');
-            upload.makePreflightRequest = jest.fn();
-            upload.retryCount = retryCount;
-
-            upload.uploadErrorHandler(error);
-            clock.runTimersToTime(2 ** retryCount * MS_IN_S - 1);
-
-            expect(upload.retryCount).toBe(retryCount + 1);
-
-            expect(upload.makePreflightRequest).not.toHaveBeenCalled();
         });
     });
 
@@ -254,8 +99,8 @@ describe('api/PlainUpload', () => {
 
         test('should make preflight request to upload file version url if fileId is given', () => {
             const baseUrl = 'base';
-            const fileId = '234';
 
+            upload.fileId = '234';
             upload.isDestroyed = jest.fn().mockReturnValueOnce(false);
             upload.getBaseApiUrl = jest.fn().mockReturnValueOnce(baseUrl);
             upload.file = {
@@ -268,11 +113,9 @@ describe('api/PlainUpload', () => {
                 options: jest.fn()
             };
 
-            upload.makePreflightRequest({
-                fileId
-            });
+            upload.makePreflightRequest();
             expect(upload.xhr.options).toHaveBeenCalledWith({
-                url: `${baseUrl}/files/${fileId}/content`,
+                url: `${baseUrl}/files/${upload.fileId}/content`,
                 data: expect.any(Object),
                 successHandler: expect.any(Function),
                 errorHandler: expect.any(Function)
@@ -280,20 +123,17 @@ describe('api/PlainUpload', () => {
         });
     });
 
-    describe('makeUploadRequest', () => {
-        test('should not do anything if API is destroyed', async () => {
+    describe('preflightSuccessHandler', () => {
+        test('should not do anything if API is destroyed', () => {
             upload.isDestroyed = jest.fn().mockReturnValueOnce(true);
             upload.xhr = {
                 uploadFile: jest.fn()
             };
-            await upload.makeUploadRequest({
-                fileId: '123',
-                fileName: 'hunter'
-            });
+            upload.preflightSuccessHandler({ data: {} });
             expect(upload.xhr.uploadFile).not.toHaveBeenCalled();
         });
 
-        test('should generate upload URL and make request if no URL is provided', async () => {
+        test('should generate upload URL and make request if no URL is provided', () => {
             upload.isDestroyed = jest.fn().mockReturnValueOnce(false);
 
             upload.file = {
@@ -304,7 +144,7 @@ describe('api/PlainUpload', () => {
                 uploadFile: jest.fn()
             };
 
-            upload.makeUploadRequest({ data: {} });
+            upload.preflightSuccessHandler({ data: {} });
             expect(upload.xhr.uploadFile).toHaveBeenCalledWith({
                 url: `${upload.uploadHost}/api/2.0/files/content`,
                 data: {
@@ -317,7 +157,7 @@ describe('api/PlainUpload', () => {
             });
         });
 
-        test('should upload to new file version if file ID is provided', async () => {
+        test('should upload to new file version if file ID is provided', () => {
             const fileId = '123';
 
             upload.isDestroyed = jest.fn().mockReturnValueOnce(false);
@@ -330,7 +170,7 @@ describe('api/PlainUpload', () => {
                 uploadFile: jest.fn()
             };
 
-            upload.makeUploadRequest({ data: {} });
+            upload.preflightSuccessHandler({ data: {} });
             expect(upload.xhr.uploadFile).toHaveBeenCalledWith({
                 url: `${upload.uploadHost}/api/2.0/files/${fileId}/content`,
                 data: expect.any(Object),
