+++ test/uploader-no-react.html
@@ -36,6 +36,7 @@
         </style>
         <link rel="stylesheet" type="text/css" href="../dev/en-US/uploader.css" />
     </head>
+
     <body>
         <div class="inputs">
             <div>
